pyproject.toml
[project]
name = "email-sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent>=0.1.3",
    "dotenv>=0.9.9",
    "openai-agents>=0.12.5",
    "pydantic[email]>=2.12.5",
    "pypdf2>=3.0.1",
    "quart>=0.20.0",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
]
